with rent

+++ CarRental/CarRental/Rent.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using CarRental.Entities;
 using CarRental.Repositories;
 
@@ -15,6 +11,6 @@ namespace CarRental
         public DateTime StartRent { get; set; }
         public DateTime EndRent { get; set; }
         public int DayRentCount { get; set; }
-        public int Fine { get; set; }
+        public double Fine { get; set; }
     }
 }
